requirements.txt
requests
Pillow
instagrapi
schedule
python-dotenv
